separating special type processors
