+++ masquerade-core/src/main/java/com/flipkart/masquerade/Masquerade.java
@@ -17,10 +17,10 @@
 package com.flipkart.masquerade;
 
 import com.flipkart.masquerade.processor.*;
+import com.flipkart.masquerade.processor.type.NoOpInitializationProcessor;
+import com.flipkart.masquerade.processor.type.ToStringInitializationProcessor;
 import com.flipkart.masquerade.rule.Rule;
-import com.flipkart.masquerade.util.Helper;
 import com.flipkart.masquerade.util.TypeSpecContainer;
-import com.flipkart.masquerade.util.Verifier;
 import com.google.common.reflect.ClassPath;
 import com.squareup.javapoet.AnnotationSpec;
 import com.squareup.javapoet.CodeBlock;
