+++ masquerade-core/src/main/java/com/flipkart/masquerade/processor/RuleProcessor.java
@@ -17,6 +17,9 @@
 package com.flipkart.masquerade.processor;
 
 import com.flipkart.masquerade.Configuration;
+import com.flipkart.masquerade.processor.type.EnumOverrideProcessor;
+import com.flipkart.masquerade.processor.type.NoOpOverrideProcessor;
+import com.flipkart.masquerade.processor.type.ToStringProcessor;
 import com.flipkart.masquerade.rule.Rule;
 import com.flipkart.masquerade.util.TypeSpecContainer;
 import com.flipkart.masquerade.util.Verifier;
